+++ test/state/tower_test.cpp
@@ -14,13 +14,13 @@ namespace test {
 #pragma GCC diagnostic ignored "-Wunused-parameter"
 // Mock function for blasting enemy actors
 void blast_enemies(PlayerId player_id, ActorId actor_id, DoubleVec2D pos,
-                   size_t blast_range) {}
+                   uint64_t blast_range) {}
 
 #pragma GCC diagnostic pop
 class TowerTest : public Test {
   protected:
     unique_ptr<Tower> tower;
-    size_t hp, max_hp, damage_points, blast_range;
+    uint64_t hp, max_hp, damage_points, blast_range;
 
     BlastCallback blast_callback;
